# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Recipez API
  description: |
    REST API for the Recipez recipe management application.

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Authorization Scopes
    Endpoints require specific scopes in the JWT token. Regular users have scopes prefixed with resource types
    (e.g., `recipe:create`, `category:read`). System user (yeschef) has additional administrative scopes.

    ## System User APIs
    Endpoints marked with `[System User]` are restricted to the system user (yeschef) and are used for
    authentication flows and user management. These require system-level scopes.
  version: 1.0.0
  contact:
    name: Recipez Support
  license:
    name: Proprietary
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/image/create:
    post:
      tags:
        - Image
      summary: Upload an image
      description: |
        Upload a base64-encoded image. Requires `image:create` scope.
        Image is saved to the uploads directory and a database record is created.
      operationId: createImage
      security:
        - bearerAuth: [image:create]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImageRequest'
      responses:
        '200':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/image/delete/{imageId}:
    delete:
      tags:
        - Image
      summary: Delete an image
      description: |
        Delete an image. Requires `image:delete` scope.
        Default system images are protected and cannot be deleted.
      operationId: deleteImage
      security:
        - bearerAuth: [image:delete]
      parameters:
        - $ref: '#/components/parameters/ImageId'
      responses:
        '200':
          description: Image deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== AI ENDPOINTS ====================
...
  /api/email/code:
    post:
      tags:
        - Email
      summary: Send verification code email
      description: |
        Send an email with a verification code. **[System User Only]**
        Requires `email:code:create` scope (system user).
      operationId: sendEmailCode
      security:
        - bearerAuth: [email:code:create]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailCodeRequest'
      responses:
        '200':
          description: Verification code email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/email/invite:
    post:
      tags:
        - Email
      summary: Send invitation email
      description: |
        Send an invitation email to join Recipez.
        Requires `email:invite:send` scope.
      operationId: sendEmailInvite
      security:
        - bearerAuth: [email:invite:send]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailInviteRequest'
      responses:
        '200':
          description: Invitation email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/email/recipe-share:
    post:
      tags:
        - Email
      summary: Send recipe link email
      description: |
        Send an email sharing a link to a recipe.
        Requires `email:recipe:share` scope.
      operationId: sendRecipeShareLink
      security:
        - bearerAuth: [email:recipe:share]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRecipeShareRequest'
      responses:
        '200':
          description: Recipe sharing email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/email/recipe-share-full:
    post:
      tags:
        - Email
      summary: Send full recipe content email
      description: |
        Send an email with the complete recipe content (ingredients, steps).
        Requires `email:recipe:share` scope.
      operationId: sendRecipeShareFull
      security:
        - bearerAuth: [email:recipe:share]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRecipeShareFullRequest'
      responses:
        '200':
          description: Full recipe email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmailSentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== CODE ENDPOINTS (SYSTEM USER) ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with scopes. Include in Authorization header as: `Bearer <token>`

        Token contains:
        - `sub`: User subject ID
        - `scope`: Space-separated list of permission scopes
        - `iat`: Issued at timestamp
        - `exp`: Expiration time (12 hours from issue)
        - `iss`: Token issuer
        - `aud`: Token audience
  parameters:
...
    ImageId:
      name: imageId
      in: path
      required: true
      description: Image UUID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions or scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            error:
              type: string
              description: Error message
    SuccessResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            success:
              type: boolean
              example: true
...
    Image:
      type: object
      properties:
        image_id:
          type: string
          format: uuid
        image_url:
          type: string
        image_author_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    CreateImageRequest:
      type: object
      required:
        - image_data
        - image_path
        - author_id
      properties:
        image_data:
          type: string
          format: byte
          description: Base64-encoded image data
        image_path:
          type: string
          description: File path where image will be stored
        author_id:
          type: string
          format: uuid
    ImageResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            image:
              $ref: '#/components/schemas/Image'

    # ==================== AI SCHEMAS ====================
...
    EmailCodeRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: "^[a-zA-Z0-9]{4}-[a-zA-Z0-9]{4}$"
          description: Verification code in format XXXX-XXXX
    EmailInviteRequest:
      type: object
      required:
        - email
        - invite_link
        - sender_name
      properties:
        email:
          type: string
          format: email
        invite_link:
          type: string
          format: uri
        sender_name:
          type: string
          minLength: 2
          maxLength: 50
    EmailRecipeShareRequest:
      type: object
      required:
        - email
        - recipe_name
        - recipe_link
        - sender_name
      properties:
        email:
          type: string
          format: email
        recipe_name:
          type: string
          minLength: 2
          maxLength: 100
        recipe_link:
          type: string
          format: uri
        sender_name:
          type: string
          minLength: 2
          maxLength: 50
    EmailRecipeShareFullRequest:
      type: object
      required:
        - email
        - recipe_name
        - recipe_ingredients
        - recipe_steps
        - sender_name
      properties:
        email:
          type: string
          format: email
        recipe_name:
          type: string
          minLength: 2
          maxLength: 100
        recipe_description:
          type: string
          maxLength: 500
          nullable: true
        recipe_ingredients:
          type: array
          items:
            type: object
            additionalProperties:
              type: string
        recipe_steps:
          type: array
          items:
            type: string
        sender_name:
          type: string
          minLength: 2
          maxLength: 50
    EmailSentResponse:
      type: object
      properties:
        response:
          type: object
          properties:
            message:
              type: string
              example: Email sent successfully

    # ==================== CODE SCHEMAS ====================
...